# budgetwise/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BudgetTracker Dashboard API
  version: 1.0.0
  description: >
    API contract for the BudgetTracker student dashboard. Designed for a Next.js app
    where API routes live under /api. Includes endpoints for dashboard data and
    AI powered insights generated server side and returned in a strict schema.
servers:
  - url: /api

tags:
  - name: Auth
  - name: Dashboard
  - name: Data
  - name: AI

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login and receive an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /dashboard/overview:
    get:
      tags: [Dashboard]
      summary: Get all dashboard data for a month in one response
      description: >
        Returns totals, chart data, and (optionally) AI sections for the dashboard.
        This is the primary endpoint for the one page dashboard experience.
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}\-(0[1-9]|1[0-2])$'
          example: "2026-02"
        - in: query
          name: includeAi
          required: false
          schema:
            type: boolean
            default: true
          description: If false, AI sections are omitted.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /expenses:
    get:
      tags: [Data]
      summary: List expenses for a month
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}\-(0[1-9]|1[0-2])$'
          example: "2026-02"
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedExpenseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Data]
      summary: Create an expense
      description: >
        Category may be omitted. Backend may categorize and return the final category.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /expenses/{expenseId}:
    put:
      tags: [Data]
      summary: Update an expense
      parameters:
        - in: path
          name: expenseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Data]
      summary: Delete an expense
      parameters:
        - in: path
          name: expenseId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /budgets:
    get:
      tags: [Data]
      summary: Get budget for a month
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}\-(0[1-9]|1[0-2])$'
          example: "2026-02"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Budget not found for month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Data]
      summary: Create or replace budget for a month
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}\-(0[1-9]|1[0-2])$'
          example: "2026-02"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBudgetRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /ai/dashboard-insights:
    post:
      tags: [AI]
      summary: Generate AI powered dashboard insights for a month
      description: >
        Server side call to OpenAI. The response is validated and normalized into
        DashboardAiResponse. Raw model output is never returned to the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardAiRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardAiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: AI output failed schema validation after normalization attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    AuthUserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time

    CreateExpenseRequest:
      type: object
      required: [amountCents, date]
      properties:
        amountCents:
          type: integer
          minimum: 0
          description: Money stored as integer cents.
        date:
          type: string
          format: date
        category:
          type: string
          description: Optional if backend categorizes.
        merchant:
          type: string
        note:
          type: string

    UpdateExpenseRequest:
      type: object
      required: [amountCents, date, category]
      properties:
        amountCents:
          type: integer
          minimum: 0
        date:
          type: string
          format: date
        category:
          type: string
        merchant:
          type: string
        note:
          type: string

    Expense:
      type: object
      required: [id, amountCents, date, category, createdAt, updatedAt]
      properties:
        id:
          type: string
        amountCents:
          type: integer
        date:
          type: string
          format: date
        category:
          type: string
        merchant:
          type: string
        note:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PagedExpenseResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        page:
          $ref: "#/components/schemas/PageMeta"

    PageMeta:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer

    UpsertBudgetRequest:
      type: object
      required: [totalLimitCents, categoryLimits]
      properties:
        totalLimitCents:
          type: integer
          minimum: 0
        categoryLimits:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
          description: Map of category name to limit in cents.

    Budget:
      type: object
      required: [month, totalLimitCents, categoryLimits, updatedAt]
      properties:
        month:
          type: string
          description: YYYY-MM
        totalLimitCents:
          type: integer
        categoryLimits:
          type: object
          additionalProperties:
            type: integer
        updatedAt:
          type: string
          format: date-time

    DashboardOverviewResponse:
      type: object
      required: [month, totals, charts]
      properties:
        month:
          type: string
          description: YYYY-MM
        totals:
          $ref: "#/components/schemas/DashboardTotals"
        charts:
          $ref: "#/components/schemas/DashboardCharts"
        ai:
          $ref: "#/components/schemas/DashboardAiResponse"
          description: Present when includeAi is true and AI is available.

    DashboardTotals:
      type: object
      required: [totalBudgetCents, totalSpentCents, remainingCents, percentSpent]
      properties:
        totalBudgetCents:
          type: integer
          minimum: 0
        totalSpentCents:
          type: integer
          minimum: 0
        remainingCents:
          type: integer
        percentSpent:
          type: number
          format: float
          minimum: 0
          maximum: 1

    DashboardCharts:
      type: object
      required: [monthlySpendingByCategory, remainingBudgetByCategory]
      properties:
        monthlySpendingByCategory:
          type: array
          description: Data for the pie chart.
          items:
            $ref: "#/components/schemas/CategorySpendSlice"
        remainingBudgetByCategory:
          type: array
          description: Data for the remaining budget bar chart.
          items:
            $ref: "#/components/schemas/CategoryRemainingBar"

    CategorySpendSlice:
      type: object
      required: [category, spentCents, percentOfTotal]
      properties:
        category:
          type: string
        spentCents:
          type: integer
          minimum: 0
        percentOfTotal:
          type: number
          format: float
          minimum: 0
          maximum: 1

    CategoryRemainingBar:
      type: object
      required: [category, limitCents, spentCents, remainingCents, percentUsed]
      properties:
        category:
          type: string
        limitCents:
          type: integer
          minimum: 0
        spentCents:
          type: integer
          minimum: 0
        remainingCents:
          type: integer
        percentUsed:
          type: number
          format: float
          minimum: 0
          maximum: 1

    DashboardAiRequest:
      type: object
      required: [month, inputs]
      properties:
        month:
          type: string
          description: YYYY-MM
        inputs:
          $ref: "#/components/schemas/DashboardAiInputs"
        options:
          $ref: "#/components/schemas/DashboardAiOptions"

    DashboardAiInputs:
      type: object
      required: [totals, categoryBreakdown, budgetByCategory]
      properties:
        totals:
          $ref: "#/components/schemas/DashboardTotals"
        categoryBreakdown:
          type: array
          description: Current spending by category, used for tips and alerts.
          items:
            $ref: "#/components/schemas/CategorySpendSlice"
        budgetByCategory:
          type: array
          description: Category budget status, used for reallocation suggestions.
          items:
            $ref: "#/components/schemas/CategoryRemainingBar"

    DashboardAiOptions:
      type: object
      properties:
        maxCards:
          type: integer
          minimum: 1
          maximum: 6
          default: 4
        maxActions:
          type: integer
          minimum: 1
          maximum: 6
          default: 3

    DashboardAiResponse:
      type: object
      required: [cards, comparison, recommendedActions, generatedAt]
      properties:
        cards:
          type: array
          description: The top recommendation cards shown in the purple section.
          items:
            $ref: "#/components/schemas/AiRecommendationCard"
        comparison:
          $ref: "#/components/schemas/AiBudgetComparison"
        recommendedActions:
          type: array
          description: Bullet style recommended actions.
          items:
            $ref: "#/components/schemas/AiActionItem"
        generatedAt:
          type: string
          format: date-time

    AiRecommendationCard:
      type: object
      required: [type, title, message]
      properties:
        type:
          type: string
          description: alert, onTrack, tip, reallocation
          enum: [alert, onTrack, tip, reallocation]
        title:
          type: string
        message:
          type: string
        savingsCents:
          type: integer
          minimum: 0
          description: Optional estimated savings.
        category:
          type: string
          description: Optional related category.

    AiBudgetComparison:
      type: object
      required: [items]
      properties:
        items:
          type: array
          description: Data for the bar chart comparing current spending vs AI recommended.
          items:
            $ref: "#/components/schemas/AiComparisonItem"

    AiComparisonItem:
      type: object
      required: [category, currentSpendCents, recommendedSpendCents]
      properties:
        category:
          type: string
        currentSpendCents:
          type: integer
          minimum: 0
        recommendedSpendCents:
          type: integer
          minimum: 0

    AiActionItem:
      type: object
      required: [severity, text]
      properties:
        severity:
          type: string
          enum: [info, warning, critical]
        text:
          type: string

    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"

    FieldError:
      type: object
      required: [field, issue]
      properties:
        field:
          type: string
        issue:
          type: string